Extract ObjectDetailView to a separate file
