+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetailView.styled.tsx
@@ -1,9 +1,6 @@
 import styled from "@emotion/styled";
-import Modal from "metabase/components/Modal";
-import { breakpointMinMedium } from "metabase/styled-components/theme/media-queries";
+
 import { color } from "metabase/lib/colors";
-import TableFooter from "../TableSimple/TableFooter";
-import { ObjectDetailBodyWrapper } from "./ObjectDetailBody.styled";
 
 interface ObjectDetailContainerProps {
   wide: boolean;
@@ -63,42 +60,6 @@ export const FitImage = styled.img`
   margin: 1rem auto;
 `;
 
-export const RootModal = styled(Modal)`
-  ${ObjectDetailContainer} {
-    overflow: hidden;
-    ${breakpointMinMedium} {
-      width: ${({ wide }) => (wide ? "64rem" : "48rem")};
-      max-width: 95vw;
-    }
-    max-height: 95vh;
-    width: 95vw;
-
-    border: 1px solid ${color("border")};
-    border-radius: 0.5rem;
-  }
-
-  ${ObjectDetailBodyWrapper} {
-    ${breakpointMinMedium} {
-      display: flex;
-      height: calc(80vh - 4rem);
-    }
-    height: calc(100vh - 8rem);
-  }
-
-  ${ObjectDetailsTable} {
-    ${breakpointMinMedium} {
-      max-height: calc(80vh - 4rem);
-    }
-  }
-
-  ${ObjectRelationships} {
-    ${breakpointMinMedium} {
-      flex: 0 0 33.3333%;
-      max-height: calc(80vh - 4rem);
-    }
-  }
-`;
-
 export interface ObjectRelationshipContentProps {
   isClickable: boolean;
 }
@@ -115,9 +76,4 @@ export const ObjectRelationContent = styled.div<ObjectRelationshipContentProps>`
   &:hover {
     color: ${props => props.isClickable && color("brand")};
   }
-`;
-
-export const PaginationFooter = styled(TableFooter)`
-  margin-top: 0.5rem;
-  text-align: right;
 `;
